requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
decorator==5.1.1
Deprecated==1.2.15
fabric==3.2.2
Faker==33.3.1
invoke==2.2.0
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
typing_extensions==4.12.2
wrapt==1.17.2
